ExecutionEngineService refactor sketch
