+++ eth-reference-tests/src/referenceTest/java/tech/pegasys/teku/reference/ManualReferenceTestRunner.java
@@ -59,6 +59,7 @@ public class ManualReferenceTestRunner extends Eth2ReferenceTestCase {
             testDefinition ->
                 SPEC.isBlank() || testDefinition.getConfigName().equalsIgnoreCase(SPEC))
         .filter(testDefinition -> testDefinition.getTestType().startsWith(TEST_TYPE))
+        .filter(testDefinition -> testDefinition.getFork().equals("merge"))
         .map(testDefinition -> Arguments.of(testDefinition.getDisplayName(), testDefinition));
   }
 }
