+++ services/powchain/src/main/java/tech/pegasys/teku/services/powchain/execution/client/ExecutionEngineClient.java
@@ -11,7 +11,7 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.executionengine.client;
+package tech.pegasys.teku.services.powchain.execution.client;
 
 import java.util.Arrays;
 import java.util.Optional;
@@ -20,11 +20,11 @@ import org.apache.tuweni.bytes.Bytes32;
 import org.web3j.protocol.core.methods.response.EthBlock;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
-import tech.pegasys.teku.spec.executionengine.client.schema.AssembleBlockRequest;
-import tech.pegasys.teku.spec.executionengine.client.schema.ExecutionPayload;
-import tech.pegasys.teku.spec.executionengine.client.schema.GenericResponse;
-import tech.pegasys.teku.spec.executionengine.client.schema.NewBlockResponse;
-import tech.pegasys.teku.spec.executionengine.client.schema.Response;
+import tech.pegasys.teku.services.powchain.execution.client.schema.AssembleBlockRequest;
+import tech.pegasys.teku.services.powchain.execution.client.schema.ExecutionPayload;
+import tech.pegasys.teku.services.powchain.execution.client.schema.GenericResponse;
+import tech.pegasys.teku.services.powchain.execution.client.schema.NewBlockResponse;
+import tech.pegasys.teku.services.powchain.execution.client.schema.Response;
 import tech.pegasys.teku.ssz.type.Bytes20;
 
 public interface ExecutionEngineClient {
