+++ services/powchain/src/main/java/tech/pegasys/teku/services/powchain/execution/client/serializer/Bytes20Deserializer.java
@@ -11,20 +11,20 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.executionengine.client.serializer;
+package tech.pegasys.teku.services.powchain.execution.client.serializer;
 
 import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.DeserializationContext;
 import com.fasterxml.jackson.databind.JsonDeserializer;
 import java.io.IOException;
-import tech.pegasys.teku.infrastructure.unsigned.UInt64;
+import tech.pegasys.teku.ssz.type.Bytes20;
 
-public class UInt64Deserializer extends JsonDeserializer<UInt64> {
+public class Bytes20Deserializer extends JsonDeserializer<Bytes20> {
 
   @Override
-  public UInt64 deserialize(JsonParser p, DeserializationContext ctxt)
+  public Bytes20 deserialize(JsonParser p, DeserializationContext ctxt)
       throws IOException, JsonProcessingException {
-    return UInt64.valueOf(p.getValueAsString());
+    return Bytes20.fromHexString(p.getValueAsString());
   }
 }
