+++ services/powchain/src/main/java/tech/pegasys/teku/services/powchain/execution/client/serializer/UInt64AsHexDeserializer.java
@@ -11,18 +11,21 @@
  * specific language governing permissions and limitations under the License.
  */
 
-package tech.pegasys.teku.spec.executionengine.client.serializer;
+package tech.pegasys.teku.services.powchain.execution.client.serializer;
 
-import com.fasterxml.jackson.core.JsonGenerator;
-import com.fasterxml.jackson.databind.JsonSerializer;
-import com.fasterxml.jackson.databind.SerializerProvider;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
 import java.io.IOException;
+import org.apache.tuweni.bytes.Bytes;
 import tech.pegasys.teku.infrastructure.unsigned.UInt64;
 
-public class UInt64Serializer extends JsonSerializer<UInt64> {
+public class UInt64AsHexDeserializer extends JsonDeserializer<UInt64> {
+
   @Override
-  public void serialize(UInt64 value, JsonGenerator gen, SerializerProvider serializers)
-      throws IOException {
-    gen.writeString(value.toString());
+  public UInt64 deserialize(JsonParser p, DeserializationContext ctxt)
+      throws IOException, JsonProcessingException {
+    return UInt64.valueOf(Bytes.fromHexStringLenient(p.getValueAsString()).toUnsignedBigInteger());
   }
 }
